+++ js/data.js
@@ -1,3 +1,13 @@
+import {getRandomIntInclusive, getRandomFloat,arrayRandElement, zeroPad, arrayRandNumber,arrayRandFeatures,arrayRandCheckout,arrayRandCheckin,arrayRandPhotos, arrayRandDescription,arrayRandTitle, unique,featuresArray} from './util.js'
+
+let typeArray = ['place', 'flat', 'house', 'bungalow'];
+let numbersArray = [1, 2, 3, 4, 5, 6, 7, 8];
+let checkinArray = ['12:00', '13:00', '14:00'];
+let checkoutArray = ['12:00', '13:00', '14:00'];
+let photosArray = ['http://o0.github.io/assets/images/tokyo/hotel1.jpg', 'http://o0.github.io/assets/images/tokyo/hotel2.jpg', 'http://o0.github.io/assets/images/tokyo/hotel3.jpg'];
+let descriptionArray = ['The best hotel room', 'It could be bettet', 'The worst hotel number', 'Really bad', 'It is not our lvl'];
+let titleArray = ['New York', 'Monaco', 'Minsk', 'Rome', 'Barcelona', 'Madrid'];
+
 const similarFeaturesCount = 10;
 const createFeatures = () => {
 
@@ -46,4 +56,4 @@ createFeatures());
 
 console.log( similarFeatures);
 
-export{similarFeaturesCount,createFeatures, similarFeatures};
+export{similarFeaturesCount,createFeatures, similarFeatures,typeArray,numbersArray,checkinArray,checkoutArray,photosArray,descriptionArray, titleArray};
